+++ ortto_flutter_sdk/lib/ortto_flutter_sdk.dart
@@ -84,8 +84,25 @@ class Ortto {
     return _platform.trackLinkClick(link);
   }
 
-  Future<bool> onbackgroundMessageReceived(Map<String, dynamic> message, {bool handleNotificationTrigger = true}) {
-    return _platform.onMessageReceived(message, handleNotificationTrigger: handleNotificationTrigger);
+  Future<bool> onBackgroundMessageReceived(
+    Map<String, dynamic> message, {
+    bool? handleNotificationTrigger,
+  }) {
+    return _platform.onMessageReceived(
+      message,
+      handleNotificationTrigger: handleNotificationTrigger ?? message['notification'] == null,
+    );
+  }
+
+  @Deprecated('Use onBackgroundMessageReceived instead.')
+  Future<bool> onbackgroundMessageReceived(
+    Map<String, dynamic> message, {
+    bool? handleNotificationTrigger,
+  }) {
+    return onBackgroundMessageReceived(
+      message,
+      handleNotificationTrigger: handleNotificationTrigger,
+    );
   }
 
   Future<IdentityResult> clearIdentity() {